+++ packages/twenty-front/src/modules/settings/accounts/components/SettingsAccountsMessageChannelDetails.tsx
@@ -9,7 +9,7 @@ import { CoreObjectNameSingular } from '@/object-metadata/types/CoreObjectNameSi
 import { useUpdateOneRecord } from '@/object-record/hooks/useUpdateOneRecord';
 import { SettingsAccountsMessageAutoCreationCard } from '@/settings/accounts/components/SettingsAccountsMessageAutoCreationCard';
 import { SettingsAccountsMessageVisibilityCard } from '@/settings/accounts/components/SettingsAccountsMessageVisibilityCard';
-import { SettingsOptionCardContentWithToggle } from '@/settings/components/SettingsOptionCardContentWithToggle';
+import { SettingsOptionCardContent } from '@/settings/components/SettingsOptionCardContent';
 import { MessageChannelVisibility } from '~/generated-metadata/graphql';
 
 type SettingsAccountsMessageChannelDetailsProps = {
@@ -99,7 +99,7 @@ export const SettingsAccountsMessageChannelDetails = ({
       </Section>
       <Section>
         <Card>
-          <SettingsOptionCardContentWithToggle
+          <SettingsOptionCardContent
             title="Exclude non-professional emails"
             description="Don’t create contacts from/to Gmail, Outlook emails"
             divider
@@ -110,7 +110,7 @@ export const SettingsAccountsMessageChannelDetails = ({
               );
             }}
           />
-          <SettingsOptionCardContentWithToggle
+          <SettingsOptionCardContent
             title="Exclude group emails"
             description="Don’t sync emails from team@ support@ noreply@..."
             checked={messageChannel.excludeGroupEmails}
